Update dependencies, replace embedded kotlin and replace authentication
# Update dependencies
One thing to note: `v3-rev20221108-2.0.0` of `google-api-services-androidpublisher` uses version `1.11` of `commons-codec:commons-codec` which has a security vulnerability: [Cxeb68d52e-5509](https://devhub.checkmarx.com/cve-details/Cxeb68d52e-5509/). We thus manually override the used version with `1.15` for now.

# Embedded kotlin replacement (`kotlin-dsl`):
One major downside of using the embedded kotlin version of gradle (by using the `kotlin-dsl` notation in the plugins block) is that one is limited to the often times pretty old kotlin version that is shipped with gradle.
As the kotlin version should be mostly independent of the gradle version used, the `kotlin-dsl` notation is replaced by the regular `kotlin("jvm")` notation. This also required slight rework of the plugin configuration as some dsl functions are no longer available in the plugin class.

# Replace authentication
The `GoogleCredential` from the `google-api-services-androidpublisher` library was deprecated a while ago. This now replaces it with `google-auth-library-oauth2-http` as suggested by the deprecation note.